Add section about setup to getting-started guide
Currently none of the CocoaPods documentation references the need to issue "pod setup" before attempting to run "pod install" in a project directory. Running the latter command will fail with an error (No such file or directory - ~/.cocoapods/repos) if setup is not run first. 

I'm not sure if this is a good place for the documentation, but it needs to be mentioned somewhere visibly for new users.